requirements.txt
# github
requests
art
termcolor

# google
google-api-python-client
google-auth-oauthlib